Move things around
I'm getting ready to add a new file that will create Swagger instances,
and I thought it shouldn't go on the same level as the other classes.
