+++ elide-contrib/elide-swagger/src/com/yahoo/elide/contrib/swagger/JSONObjectClasses/Enums.java
@@ -1,4 +1,4 @@
-package com.yahoo.elide.contrib.swagger;
+package com.yahoo.elide.contrib.swagger.JSONObjectClasses;
 
 import com.google.gson.annotations.SerializedName;
