+++ elide-contrib/elide-swagger/src/com/yahoo/elide/contrib/swagger/JSONObjectClasses/SwaggerComponent.java
@@ -1,4 +1,4 @@
-package com.yahoo.elide.contrib.swagger;
+package com.yahoo.elide.contrib.swagger.JSONObjectClasses;
 
 import java.lang.reflect.Field;
 import java.util.HashMap;
